+++ packages/rich-text/src/plugins/EmbeddedEntityInline/FetchingWrappedInlineEntryCard.tsx
@@ -12,7 +12,7 @@ import { Entry, FieldAppSDK, entityHelpers } from '@contentful/field-editor-shar
 import { INLINES } from '@contentful/rich-text-types';
 import { css } from 'emotion';
 
-const { getEntryTitle, getEntryStatus } = entityHelpers;
+const { getEntryTitle, getEntityStatus } = entityHelpers;
 
 const styles = {
   scheduledIcon: css({
@@ -20,6 +20,81 @@ const styles = {
     marginRight: tokens.spacing2Xs,
   }),
 };
+
+type InternalFetchingWrappedInlineEntryCardProps = Pick<
+  FetchingWrappedInlineEntryCardProps,
+  'onEdit' | 'onRemove' | 'isDisabled' | 'isSelected'
+> & {
+  locale: string;
+  defaultLocale: string;
+  entry: Entry;
+  allContentTypes: ReturnType<FieldAppSDK['space']['getCachedContentTypes']>;
+  getEntityScheduledActions: React.ComponentProps<
+    typeof ScheduledIconWithTooltip
+  >['getEntityScheduledActions'];
+  entryStatus: ReturnType<typeof getEntityStatus>;
+};
+
+function InternalFetchingWrappedInlineEntryCard({
+  entry,
+  allContentTypes,
+  locale,
+  defaultLocale,
+  isSelected,
+  entryStatus,
+  getEntityScheduledActions,
+  onEdit,
+  onRemove,
+  isDisabled,
+}: InternalFetchingWrappedInlineEntryCardProps) {
+  const contentType = React.useMemo(() => {
+    if (!allContentTypes) {
+      return undefined;
+    }
+
+    return allContentTypes.find(
+      (contentType) => contentType.sys.id === entry.sys.contentType.sys.id
+    );
+  }, [allContentTypes, entry]);
+
+  const title = React.useMemo(
+    () =>
+      getEntryTitle({
+        entry,
+        contentType,
+        localeCode: locale,
+        defaultLocaleCode: defaultLocale,
+        defaultTitle: 'Untitled',
+      }),
+    [entry, contentType, locale, defaultLocale]
+  );
+
+  return (
+    <InlineEntryCard
+      testId={INLINES.EMBEDDED_ENTRY}
+      isSelected={isSelected}
+      title={contentType ? `${contentType.name}: ${title}` : title}
+      status={entryStatus}
+      actions={[
+        <MenuItem key="edit" onClick={onEdit}>
+          Edit
+        </MenuItem>,
+        <MenuItem key="remove" onClick={onRemove} disabled={isDisabled} testId="delete">
+          Remove
+        </MenuItem>,
+      ]}
+    >
+      <ScheduledIconWithTooltip
+        getEntityScheduledActions={getEntityScheduledActions}
+        entityType="Entry"
+        entityId={entry.sys.id}
+      >
+        <ClockIcon className={styles.scheduledIcon} variant="muted" testId="scheduled-icon" />
+      </ScheduledIconWithTooltip>
+      <Text>{title}</Text>
+    </InlineEntryCard>
+  );
+}
 
 interface FetchingWrappedInlineEntryCardProps {
   entryId: string;
@@ -34,17 +109,8 @@ interface FetchingWrappedInlineEntryCardProps {
 export function FetchingWrappedInlineEntryCard(props: FetchingWrappedInlineEntryCardProps) {
   const { data: entry, status: requestStatus } = useEntity<Entry>('Entry', props.entryId);
   const { getEntityScheduledActions } = useEntityLoader();
-  const loadEntityScheduledActions = () => getEntityScheduledActions('Entry', props.entryId);
 
-  const allContentTypes = props.sdk.space.getCachedContentTypes();
   const { onEntityFetchComplete } = props;
-  const contentType = React.useMemo(() => {
-    if (!entry || !allContentTypes) return undefined;
-
-    return allContentTypes.find(
-      (contentType) => contentType.sys.id === entry.sys.contentType.sys.id
-    );
-  }, [allContentTypes, entry]);
 
   React.useEffect(() => {
     if (requestStatus !== 'success') {
@@ -53,39 +119,29 @@ export function FetchingWrappedInlineEntryCard(props: FetchingWrappedInlineEntry
     onEntityFetchComplete?.();
   }, [requestStatus, onEntityFetchComplete]);
 
-  const contentTypeName = contentType ? contentType.name : '';
-
-  const title = React.useMemo(
-    () =>
-      getEntryTitle({
-        entry,
-        contentType,
-        localeCode: props.sdk.field.locale,
-        defaultLocaleCode: props.sdk.locales.default,
-        defaultTitle: 'Untitled',
-      }),
-    [entry, contentType, props.sdk.field.locale, props.sdk.locales.default]
-  );
+  if (requestStatus === 'loading' || requestStatus === 'idle') {
+    return <InlineEntryCard isLoading />;
+  }
 
   if (requestStatus === 'error') {
     return (
       <InlineEntryCard
-        title="Entry missing or inaccessible"
+        title="Content missing or inaccessible"
         testId={INLINES.EMBEDDED_ENTRY}
         isSelected={props.isSelected}
       />
     );
   }
 
-  if (requestStatus === 'loading') {
-    return <InlineEntryCard isLoading />;
-  }
+  const entryStatus = getEntityStatus(
+    entry.sys,
+    props.sdk.parameters.instance.useLocalizedEntityStatus ? props.sdk.field.locale : undefined
+  );
 
-  const entryStatus = getEntryStatus(entry.sys);
   if (entryStatus === 'deleted') {
     return (
       <InlineEntryCard
-        title="Entry missing or inaccessible"
+        title="Content missing or inaccessible"
         testId={INLINES.EMBEDDED_ENTRY}
         isSelected={props.isSelected}
         actions={[
@@ -98,28 +154,17 @@ export function FetchingWrappedInlineEntryCard(props: FetchingWrappedInlineEntry
   }
 
   return (
-    <InlineEntryCard
-      testId={INLINES.EMBEDDED_ENTRY}
+    <InternalFetchingWrappedInlineEntryCard
+      allContentTypes={props.sdk.space.getCachedContentTypes()}
+      getEntityScheduledActions={() => getEntityScheduledActions('Entry', props.entryId)}
+      locale={props.sdk.field.locale}
+      defaultLocale={props.sdk.locales.default}
+      entry={entry}
+      entryStatus={entryStatus}
+      isDisabled={props.isDisabled}
       isSelected={props.isSelected}
-      title={`${contentTypeName}: ${title}`}
-      status={entryStatus}
-      actions={[
-        <MenuItem key="edit" onClick={props.onEdit}>
-          Edit
-        </MenuItem>,
-        <MenuItem key="remove" onClick={props.onRemove} disabled={props.isDisabled} testId="delete">
-          Remove
-        </MenuItem>,
-      ]}
-    >
-      <ScheduledIconWithTooltip
-        getEntityScheduledActions={loadEntityScheduledActions}
-        entityType="Entry"
-        entityId={entry.sys.id}
-      >
-        <ClockIcon className={styles.scheduledIcon} variant="muted" testId="scheduled-icon" />
-      </ScheduledIconWithTooltip>
-      <Text>{title}</Text>
-    </InlineEntryCard>
+      onEdit={props.onEdit}
+      onRemove={props.onRemove}
+    />
   );
 }