+++ packages/rich-text/src/plugins/shared/__tests__/FetchingWrappedAssetCard.test.tsx
@@ -35,6 +35,10 @@ beforeEach(() => {
       space: 'space-id',
       environment: 'environment-id',
     },
+    parameters: {
+      instance: {},
+    },
+    field: { localized: false },
   };
 });