+++ packages/single-line/src/SingleLineEditor.tsx
@@ -12,7 +12,6 @@ import {
 
 import * as styles from './styles';
 
-
 export interface SingleLineEditorProps {
   /**
    * is the field disabled initially
@@ -37,6 +36,16 @@ export interface SingleLineEditorProps {
    * sdk.locales
    */
   locales: LocalesAPI;
+
+  /**
+   * blur event handler
+   */
+  onBlur?: (event: React.FocusEvent<HTMLInputElement>) => void;
+
+  /**
+   * focus event handler
+   */
+  onFocus?: (event: React.FocusEvent<HTMLInputElement>) => void;
 }
 
 function isSupportedFieldTypes(val: string): val is 'Symbol' | 'Text' {
@@ -58,7 +67,8 @@ export function SingleLineEditor(props: SingleLineEditorProps) {
     <FieldConnector<string>
       field={field}
       isInitiallyDisabled={props.isInitiallyDisabled}
-      isDisabled={props.isDisabled}>
+      isDisabled={props.isDisabled}
+    >
       {({ value, errors, disabled, setValue }) => {
         return (
           <div data-test-id="single-line-editor">
@@ -68,6 +78,8 @@ export function SingleLineEditor(props: SingleLineEditorProps) {
               isInvalid={errors.length > 0}
               isDisabled={disabled}
               value={value || ''}
+              onFocus={props.onFocus}
+              onBlur={props.onBlur}
               onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
                 setValue(e.target.value);
               }}